+++ src/components/datatable/BottomPanel.js
@@ -8,7 +8,7 @@ import {
     LAYERS_PANEL_WIDTH,
     RIGHT_PANEL_WIDTH,
 } from '../../constants/layout.js'
-import DataTable from '../datatable/UiDataTable.js'
+import DataTable from '../datatable/DataTable.js'
 import { useWindowDimensions } from '../WindowDimensionsProvider.js'
 import ResizeHandle from './ResizeHandle.js'
 import styles from './styles/BottomPanel.module.css'
