+++ src/components/datatable/styles/DataTable.module.css
@@ -1,9 +1,29 @@
 .dataTable {
-    font-size: 11px;
-    font-family: Roboto, sans-serif;
-    overflow: hidden;
+    height: 1px;
 }
 
+td.dataCell {
+    padding-top: var(--spacers-dp8);
+    padding-bottom: var(--spacers-dp8);
+    font-size: 12px;
+}
+
+td.lightText {
+    color: var(--colors-white);
+}
+
+.columnHeader > :global(span.container) {
+    justify-content: space-between;
+}
+/* Hide the filter icon */
+.columnHeader
+    > :global(span.container)
+    > :global(span.top)
+    > button:last-of-type {
+    visibility: hidden;
+}
+
+/* TODO */
 .noSupport {
     position: absolute;
     top: 50%;
@@ -12,63 +32,4 @@
     color: #333;
     font-style: italic;
     line-height: 30px;
-}
-
-.loader {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-}
-
-:global(.ReactVirtualized__Table__headerRow) {
-    position: relative;
-    background-color: #eee;
-    box-shadow: rgba(0, 0, 0, 0.12) 0px 1px 6px, rgba(0, 0, 0, 0.12) 0px 1px 2px;
-    z-index: 10;
-}
-
-:global(.ReactVirtualized__Table__headerRow .ColumnHeader-label) {
-    white-space: nowrap;
-}
-
-:global(.ReactVirtualized__Table__headerTruncatedText) {
-    display: inline-block;
-    position: absolute;
-    top: 7px;
-    left: 10px;
-    right: 15px;
-    user-select: none;
-    overflow: hidden;
-}
-
-:global(.ReactVirtualized__Table__sortableHeaderIcon) {
-    position: absolute;
-    top: 5px;
-    right: 0;
-}
-
-:global(.ReactVirtualized__Table__headerColumn) {
-    height: 48px;
-    margin-bottom: 2px;
-    position: relative;
-}
-
-:global(.ReactVirtualized__Table__headerColumn:focus) {
-    outline: none;
-}
-
-:global(.ReactVirtualized__Table__row) {
-    position: relative;
-    border-bottom: 1px solid #e0e0e0;
-    z-index: 9;
-}
-
-:global(.ReactVirtualized__Table__row:hover) {
-    background-color: #fafafa;
-    cursor: pointer;
-}
-
-:global(.ReactVirtualized__Table__rowColumn.right) {
-    text-align: right;
 }
