--- src/components/datatable/styles/UiDataTable.module.css
@@ -1,35 +1,0 @@
-.dataTable {
-    height: 1px;
-}
-
-td.dataCell {
-    padding-top: var(--spacers-dp8);
-    padding-bottom: var(--spacers-dp8);
-    font-size: 12px;
-}
-
-td.lightText {
-    color: var(--colors-white);
-}
-
-.columnHeader > :global(span.container) {
-    justify-content: space-between;
-}
-/* Hide the filter icon */
-.columnHeader
-    > :global(span.container)
-    > :global(span.top)
-    > button:last-of-type {
-    visibility: hidden;
-}
-
-/* TODO */
-.noSupport {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translateX(-50%) translateY(-50%);
-    color: #333;
-    font-style: italic;
-    line-height: 30px;
-}
